lz4.go: Fix Go 1.16 SliceHeader vet warnings
Go 1.16 added vet warnings to check for unsafe uses of
reflect.SliceHeader. To fix them, create an empty slice, then alter
that slice using *reflect.SliceHeader. This is apparently safer
because it works with escape analysis and garbage collection [1].
This pattern is also used in golang.org/x/sys so this should be safe.
For example, see Mmap in syscall_unix.go:

https://cs.opensource.google/go/x/sys/+/master:unix/syscall_unix.go

[1] https://github.com/golang/go/issues/40701

Fixes:

./lz4.go:162:35: possible misuse of reflect.SliceHeader
./lz4.go:509:35: possible misuse of reflect.SliceHeader