Writer: Fix bug in last commit; fix old streaming bug on Mac OS X
Commit 7f71d02eef fixed unsafe slice warnings, but incorrectly
removed a line of code from Writer.nextInputBuffer. This causes the
writer to produce incorrect results. Add a test that reproduces this
bug.

On Mac OS X, this test also often fails even with this bug fix, or
with the version before the bug. It turns out that
LZ4_compress_fast_continue will combine contiguous buffers for small
writes, in an attempt to improve compression. To disable this, we
must separate the buffers. See:

https://github.com/lz4/lz4/issues/473#issuecomment-366537441