lz4_test.go: TestDecompConcurrently: Actually run in parallel
This test appears to have intended to run a number of sub-tests in
parallel, likely to try and test concurrent decompression. However,
the call to t.Parallel() only marked the top-level test as being
able to execute in parallel. Instead, use a wait group with
goroutines to make it explicit what is being checked, and to ensure
that the code actually runs in parallel.

I tried moving the t.Parallel() call, but it was more confusing due
to needing to defer when the input file is deleted.